@@ -1,1 +1,1 @@
-2.7.6
+3.0.6
