@@ -1,31 +1,30 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby '2.7.6'
+ruby '3.0.6'
 
-gem 'rails', '~> 5.2.6'
-gem 'puma', '~> 3.11'
-gem 'sass-rails', '~> 5.0'
-gem 'uglifier', '>= 1.3.0'
-gem 'coffee-rails', '~> 4.2'
-gem 'turbolinks', '~> 5'
-gem 'jbuilder', '~> 2.5'
-gem 'bootsnap', '>= 1.1.0', require: false
-gem 'listen', '>= 3.0.5', '< 3.2'
-gem 'byebug'
+gem 'rails', '6.1.7.2'
+gem 'puma'
+gem 'sass-rails'
+gem 'uglifier'
+gem 'coffee-rails'
+gem 'turbolinks'
+gem 'jbuilder'
+gem 'bootsnap', require: false
+gem 'listen'
 gem 'dotenv-rails'
 
 # ログイン機能
 gem 'devise'
 
 # 日本語化
-gem 'rails-i18n', '~> 5.1'
+gem 'rails-i18n'
 gem 'devise-i18n'
 
 # Bootstrap
-gem 'bootstrap', '~> 4.5'
+gem 'bootstrap'
 gem 'jquery-rails'
-gem 'devise-bootstrap-views', '~> 1.0'
+gem 'devise-bootstrap-views'
 
 # ページネーション
 gem 'will_paginate'
@@ -40,27 +39,25 @@ gem 'roo'
 # グラフ
 gem 'chartkick'
 
-# gem 'sqlite3'
 group :development, :test do
   gem 'sqlite3'
   gem 'byebug', platforms: [:mri, :mingw, :x64_mingw]
-  gem 'rspec-rails', '~> 3.8.0'
-  gem "factory_bot_rails", "~> 4.10.0"
+  gem 'rspec-rails'
+  gem "factory_bot_rails"
 end
 
 group :development do
-  gem 'web-console', '>= 3.3.0'
+  gem 'web-console'
   gem 'spring'
-  gem 'spring-watcher-listen', '~> 2.0.0'
+  gem 'spring-watcher-listen'
   gem 'spring-commands-rspec'
   gem 'rubocop', require: false
 end
 
 group :test do
-  gem 'capybara', '>= 2.15'
-  gem 'launchy', '~> 2.4.3'
+  gem 'capybara'
+  gem 'launchy'
   gem 'selenium-webdriver'
-  # gem 'chromedriver-helper'
   gem 'webdrivers'
 end
 
