@@ -44,7 +44,7 @@ class Staffs::SessionsController < Devise::SessionsController
       respond_with resource, location: after_sign_in_path_for(resource)
     else
       flash[:danger] = "登録された名前が違います"
-      redirect_to root_path
+      redirect_to new_staff_session_path
     end
   end
 
