@@ -2,11 +2,11 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
@@ -128,4 +128,10 @@ ActiveRecord::Schema.define(version: 2021_12_05_125430) do
     t.index ["reset_password_token"], name: "index_voters_on_reset_password_token", unique: true
   end
 
+  add_foreign_key "adminmessages", "staffs"
+  add_foreign_key "group_email_comments", "staffs"
+  add_foreign_key "likes", "posts"
+  add_foreign_key "likes", "voters"
+  add_foreign_key "posts", "users"
+  add_foreign_key "voterposts", "voters"
 end
