@@ -4,7 +4,7 @@ FactoryBot.define do
     sequence(:second_phrase) { |n| "かきくけこさし#{n}" }
     sequence(:third_phrase) { |n| "たちつてと#{n}" }
     sequence(:pen_name) { |n| "post_name#{n}" }
-    category "女性パワーで未来を変えるの部"
+    category { "女性パワーで未来を変えるの部" }
     sequence(:user_id) { |n| n }
     association :user
   end
