@@ -2,8 +2,8 @@ FactoryBot.define do
   factory :voter do
     sequence(:name) { |n| "test-voter#{n}" }
     sequence(:email) { |n| "test-voter#{n}@email.com" }
-    group "WIP"
-    password "password"
+    group { "WIP" }
+    password { "password" }
   end
 
   trait :with_likes_for_voter do
